requirements.txt
Flask==3.0.0
fastapi==0.110.0
pandas==2.2.1
openai==1.2.0
requests==2.31.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfplumber==0.10.2
psycopg2-binary==2.9.9
smtplib==3.11.0
jira==3.5.0